requirements.txt
ipython==7.13.0
jupyter==1.0.0
matplotlib==3.2.2
notebook==6.0.3
numpy==1.18.1
pandas==1.0.3
Pillow==7.1.2
spyder==3.3.3
torch==1.6.0
torchvision==0.7.0
tqdm==4.31.1
opencv-python==4.4.0.42
torchfile==0.1.0
